Optimizations for Record Grouping #D-824
Also addresses the ability to turn off suppressing based on OverDrive
and Hoopla urls #D-1623

+++ vufind/record_grouping/src/org/vufind/RecordGrouperMain.java
@@ -36,8 +36,6 @@ public class RecordGrouperMain {
 	private static String serverName;
 
 	static String groupedWorkTableName = "grouped_work";
-	static String groupedWorkIdentifiersTableName = "grouped_work_identifiers";
-	static String groupedWorkIdentifiersRefTableName = "grouped_work_identifiers_ref";
 
 	private static HashMap<String, Long> marcRecordChecksums = new HashMap<>();
 	private static HashMap<String, Long> marcRecordFirstDetectionDates = new HashMap<>();
@@ -82,7 +80,9 @@ public class RecordGrouperMain {
 					"5) Only run record grouping cleanup\n" +
 					"   record_grouping.jar <pika_site_name> runPostGroupingCleanup\n" +
 					"6) Only explode records into individual records (no grouping)\n" +
-					"   record_grouping.jar <pika_site_name> explodeMarcs");
+					"   record_grouping.jar <pika_site_name> explodeMarcs\n" +
+					"7) Record Group a specific indexing profile\n" +
+					"   record_grouping.jar <pika_site_name> \"<profile name>\"");
 			System.exit(1);
 		}
 
@@ -667,9 +667,6 @@ public class RecordGrouperMain {
 				clearDatabase(vufindConn, clearDatabasePriorToGrouping);
 			}
 
-			loadIlsChecksums(vufindConn, indexingProfileToRun);
-			loadExistingPrimaryIdentifiers(vufindConn, indexingProfileToRun);
-
 			//Determine if we want to validateChecksumsFromDisk
 			try{
 				PreparedStatement getValidateChecksumsFromDiskVariableStmt = vufindConn.prepareStatement("SELECT * FROM variables where name = 'validateChecksumsFromDisk'");
@@ -707,6 +704,7 @@ public class RecordGrouperMain {
 					profile.format = getCharFromRecordSet(indexingProfilesRS, "format");
 					profile.itemTag = indexingProfilesRS.getString("itemTag");
 					profile.eContentDescriptor = getCharFromRecordSet(indexingProfilesRS, "eContentDescriptor");
+					profile.doAutomaticEcontentSuppression = indexingProfilesRS.getBoolean("doAutomaticEcontentSuppression");
 
 					indexingProfiles.add(profile);
 				}
@@ -716,16 +714,12 @@ public class RecordGrouperMain {
 			}
 
 			if (indexingProfileToRun == null || indexingProfileToRun.equalsIgnoreCase("overdrive")) {
-				groupOverDriveRecords(configIni, econtentConnection, recordGroupingProcessor, explodeMarcsOnly);
+				groupOverDriveRecords(configIni, vufindConn, econtentConnection, recordGroupingProcessor, explodeMarcsOnly);
 			}
 			if (indexingProfiles.size() > 0) {
 				groupIlsRecords(configIni, vufindConn, indexingProfiles, explodeMarcsOnly);
 			}
 
-			//Remove deleted records now that we have processed all records that currently exist
-			if (!explodeMarcsOnly) {
-				removeDeletedRecords();
-			}
 		}
 
 		if (!explodeMarcsOnly) {
@@ -770,33 +764,39 @@ public class RecordGrouperMain {
 		logger.info("Elapsed Minutes " + (elapsedTime / 60000));
 	}
 
-	private static void removeDeletedRecords() {
-		logger.info("Deleting " + marcRecordIdsInDatabase.size() + " record ids from the database since they are no longer in the export.");
-		for (String recordNumber : marcRecordIdsInDatabase.keySet()) {
-			//Remove the record from the ils_marc_checksums table
-			try {
-				removeMarcRecordChecksum.setLong(1, marcRecordIdsInDatabase.get(recordNumber));
-				int numRemoved = removeMarcRecordChecksum.executeUpdate();
-				if (numRemoved != 1){
-					logger.warn("Could not delete " + recordNumber +  " from ils_marc_checksums table");
+	private static void removeDeletedRecords(String curProfile) {
+		if (marcRecordIdsInDatabase.size() > 0) {
+			logger.info("Deleting " + marcRecordIdsInDatabase.size() + " record ids for profile " + curProfile + " from the database since they are no longer in the export.");
+			for (String recordNumber : marcRecordIdsInDatabase.keySet()) {
+				//Remove the record from the ils_marc_checksums table
+				try {
+					removeMarcRecordChecksum.setLong(1, marcRecordIdsInDatabase.get(recordNumber));
+					int numRemoved = removeMarcRecordChecksum.executeUpdate();
+					if (numRemoved != 1) {
+						logger.warn("Could not delete " + recordNumber + " from ils_marc_checksums table");
+					}
+				} catch (SQLException e) {
+					logger.error("Error removing ILS id " + recordNumber + " from ils_marc_checksums table", e);
 				}
-			} catch (SQLException e) {
-				logger.error("Error removing ILS id " + recordNumber + " from ils_marc_checksums table", e);
 			}
+			marcRecordIdsInDatabase.clear();
 		}
 
-		logger.info("Deleting " + primaryIdentifiersInDatabase.size() + " primary identifiers from the database since they are no longer in the export.");
-		for (String recordNumber : primaryIdentifiersInDatabase.keySet()) {
-			//Remove the record from the grouped_work_primary_identifiers table
-			try {
-				removePrimaryIdentifier.setLong(1, primaryIdentifiersInDatabase.get(recordNumber));
-				int numRemoved = removePrimaryIdentifier.executeUpdate();
-				if (numRemoved != 1){
-					logger.warn("Could not delete " + recordNumber +  " from grouped_work_primary_identifiers table");
+		if (primaryIdentifiersInDatabase.size() > 0) {
+			logger.info("Deleting " + primaryIdentifiersInDatabase.size() + " primary identifiers for profile " + curProfile + " from the database since they are no longer in the export.");
+			for (String recordNumber : primaryIdentifiersInDatabase.keySet()) {
+				//Remove the record from the grouped_work_primary_identifiers table
+				try {
+					removePrimaryIdentifier.setLong(1, primaryIdentifiersInDatabase.get(recordNumber));
+					int numRemoved = removePrimaryIdentifier.executeUpdate();
+					if (numRemoved != 1) {
+						logger.warn("Could not delete " + recordNumber + " from grouped_work_primary_identifiers table");
+					}
+				} catch (SQLException e) {
+					logger.error("Error removing " + recordNumber + " from grouped_work_primary_identifiers table", e);
 				}
-			} catch (SQLException e) {
-				logger.error("Error removing " + recordNumber + " from grouped_work_primary_identifiers table", e);
 			}
+			primaryIdentifiersInDatabase.clear();
 		}
 	}
 
@@ -1040,8 +1040,10 @@ public class RecordGrouperMain {
 	private static void loadExistingPrimaryIdentifiers(Connection vufindConn, String indexingProfileToRun) {
 		//Load MARC Existing MARC Record checksums from VuFind
 		try{
-			insertMarcRecordChecksum = vufindConn.prepareStatement("INSERT INTO ils_marc_checksums (ilsId, source, checksum, dateFirstDetected) VALUES (?, ?, ?, ?) ON DUPLICATE KEY UPDATE checksum = VALUES(checksum), dateFirstDetected=VALUES(dateFirstDetected), source=VALUES(source)");
-			removeMarcRecordChecksum = vufindConn.prepareStatement("DELETE FROM ils_marc_checksums WHERE id = ?");
+			if (insertMarcRecordChecksum == null) {
+				insertMarcRecordChecksum = vufindConn.prepareStatement("INSERT INTO ils_marc_checksums (ilsId, source, checksum, dateFirstDetected) VALUES (?, ?, ?, ?) ON DUPLICATE KEY UPDATE checksum = VALUES(checksum), dateFirstDetected=VALUES(dateFirstDetected), source=VALUES(source)");
+				removeMarcRecordChecksum = vufindConn.prepareStatement("DELETE FROM ils_marc_checksums WHERE id = ?");
+			}
 
 			//MDN 2/23/2015 - Always load checksums so we can optimize writing to the database
 			PreparedStatement loadIlsMarcChecksums;
@@ -1052,9 +1054,10 @@ public class RecordGrouperMain {
 				loadIlsMarcChecksums.setString(1, indexingProfileToRun);
 			}
 			ResultSet ilsMarcChecksumRS = loadIlsMarcChecksums.executeQuery();
+			Long zero = 0L;
 			while (ilsMarcChecksumRS.next()){
 				Long checksum = ilsMarcChecksumRS.getLong("checksum");
-				if (checksum == 0){
+				if (checksum.equals(zero)){
 					checksum = null;
 				}
 				String fullIdentifier = ilsMarcChecksumRS.getString("source") + ":" + ilsMarcChecksumRS.getString("ilsId").trim();
@@ -1079,8 +1082,10 @@ public class RecordGrouperMain {
 
 	private static void loadIlsChecksums(Connection vufindConn, String indexingProfileToRun) {
 		//Load Existing Primary Identifiers so we can clean up
-		try{
-			removePrimaryIdentifier = vufindConn.prepareStatement("DELETE FROM grouped_work_primary_identifiers WHERE id = ?");
+		try {
+			if (removePrimaryIdentifier == null){
+				removePrimaryIdentifier = vufindConn.prepareStatement("DELETE FROM grouped_work_primary_identifiers WHERE id = ?");
+			}
 
 			PreparedStatement loadPrimaryIdentifiers;
 			if (indexingProfileToRun == null) {
@@ -1111,7 +1116,9 @@ public class RecordGrouperMain {
 			try{
 				vufindConn.prepareStatement("TRUNCATE ils_marc_checksums").executeUpdate();
 				vufindConn.prepareStatement("TRUNCATE " + groupedWorkTableName).executeUpdate();
+				String groupedWorkIdentifiersTableName = "grouped_work_identifiers";
 				vufindConn.prepareStatement("TRUNCATE " + groupedWorkIdentifiersTableName).executeUpdate();
+				String groupedWorkIdentifiersRefTableName = "grouped_work_identifiers_ref";
 				vufindConn.prepareStatement("TRUNCATE " + groupedWorkIdentifiersRefTableName).executeUpdate();
 				String groupedWorkPrimaryIdentifiersTableName = "grouped_work_primary_identifiers";
 				vufindConn.prepareStatement("TRUNCATE " + groupedWorkPrimaryIdentifiersTableName).executeUpdate();
@@ -1126,34 +1133,41 @@ public class RecordGrouperMain {
 	private static void groupIlsRecords(Ini configIni, Connection dbConnection, ArrayList<IndexingProfile> indexingProfiles, boolean explodeMarcsOnly) {
 		//Get indexing profiles
 		for (IndexingProfile curProfile : indexingProfiles) {
+			loadIlsChecksums(dbConnection, curProfile.name);
+			loadExistingPrimaryIdentifiers(dbConnection, curProfile.name);
+
 			MarcRecordGrouper recordGroupingProcessor;
-			if (curProfile.groupingClass.equals("MarcRecordGrouper")) {
-				recordGroupingProcessor = new MarcRecordGrouper(dbConnection, curProfile, logger, fullRegrouping);
-			}else if (curProfile.groupingClass.equals("SideLoadedRecordGrouper")){
-				recordGroupingProcessor = new SideLoadedRecordGrouper(dbConnection, curProfile, logger, fullRegrouping);
-			}else if (curProfile.groupingClass.equals("HooplaRecordGrouper")){
-				recordGroupingProcessor = new HooplaRecordGrouper(dbConnection, curProfile, logger, fullRegrouping);
-			}else{
-				logger.error("Unknown class for record grouping " + curProfile.groupingClass);
-				continue;
+			switch (curProfile.groupingClass) {
+				case "MarcRecordGrouper":
+					recordGroupingProcessor = new MarcRecordGrouper(dbConnection, curProfile, logger, fullRegrouping);
+					break;
+				case "SideLoadedRecordGrouper":
+					recordGroupingProcessor = new SideLoadedRecordGrouper(dbConnection, curProfile, logger, fullRegrouping);
+					break;
+				case "HooplaRecordGrouper":
+					recordGroupingProcessor = new HooplaRecordGrouper(dbConnection, curProfile, logger, fullRegrouping);
+					break;
+				default:
+					logger.error("Unknown class for record grouping " + curProfile.groupingClass);
+					continue;
 			}
 
 			logger.debug("Processing profile " + curProfile.name);
 
 			int numRecordsProcessed = 0;
 			int numRecordsRead = 0;
-			String individualMarcPath = curProfile.individualMarcPath;
 			String marcPath = curProfile.marcPath;
 
 			String marcEncoding = curProfile.marcEncoding;
 
 			//Load all files in the individual marc path.  This allows us to list directories rather than doing millions of
 			//individual look ups
+			/*String individualMarcPath = curProfile.individualMarcPath;
 			HashSet<String> existingMarcFiles = new HashSet<>();
 			File individualMarcFile = new File(individualMarcPath);
 			logger.debug("Starting to read existing marc files for ILS from disc");
 			loadExistingMarcFiles(individualMarcFile, existingMarcFiles);
-			logger.debug("Finished reading existing marc files for ILS from disc");
+			logger.debug("Finished reading existing marc files for ILS from disc");*/
 
 			TreeSet<String> recordNumbersInExport = new TreeSet<>();
 			TreeSet<String> suppressedRecordNumbersInExport = new TreeSet<>();
@@ -1174,7 +1188,7 @@ public class RecordGrouperMain {
 							while (catalogReader.hasNext()) {
 								try{
 									Record curBib = catalogReader.next();
-									RecordIdentifier recordIdentifier = recordGroupingProcessor.getPrimaryIdentifierFromMarcRecord(curBib, curProfile.name);
+									RecordIdentifier recordIdentifier = recordGroupingProcessor.getPrimaryIdentifierFromMarcRecord(curBib, curProfile.name, curProfile.doAutomaticEcontentSuppression);
 									if (recordIdentifier == null) {
 										//logger.debug("Record with control number " + curBib.getControlNumber() + " was suppressed or is eContent");
 										String controlNumber = curBib.getControlNumber();
@@ -1189,7 +1203,7 @@ public class RecordGrouperMain {
 									}else{
 										String recordNumber = recordIdentifier.getIdentifier();
 
-										boolean marcUpToDate = writeIndividualMarc(existingMarcFiles, curProfile, curBib, recordNumber, marcRecordsWritten, marcRecordsOverwritten);
+										boolean marcUpToDate = writeIndividualMarc(curProfile, curBib, recordNumber, marcRecordsWritten, marcRecordsOverwritten);
 										recordNumbersInExport.add(recordIdentifier.toString());
 										if (!explodeMarcsOnly) {
 											if (!marcUpToDate || fullRegroupingNoClear) {
@@ -1231,17 +1245,29 @@ public class RecordGrouperMain {
 				}
 			}
 
+			removeDeletedRecords(curProfile.name);
+
 			String profileName = curProfile.name.replaceAll(" ", "_");
 			writeExistingRecordsFile(configIni, recordNumbersInExport, "record_grouping_" + profileName + "_bibs_in_export");
-			writeExistingRecordsFile(configIni, suppressedRecordNumbersInExport, "record_grouping_" + profileName + "_bibs_to_ignore");
-			writeExistingRecordsFile(configIni, suppressedRecordNumbersInExport, "record_grouping_" + profileName + "_ccontrol_numbers_to_ignore");
-			writeExistingRecordsFile(configIni, recordNumbersToIndex, "record_grouping_" + profileName + "_bibs_to_index");
-			writeExistingRecordsFile(configIni, marcRecordsWritten, "record_grouping_" + profileName + "_new_bibs_written");
-			writeExistingRecordsFile(configIni, marcRecordsOverwritten, "record_grouping_" + profileName + "_changed_bibs_written");
+			if (suppressedRecordNumbersInExport.size() > 0) {
+				writeExistingRecordsFile(configIni, suppressedRecordNumbersInExport, "record_grouping_" + profileName + "_bibs_to_ignore");
+			}
+			if (suppressedControlNumbersInExport.size() > 0) {
+				writeExistingRecordsFile(configIni, suppressedControlNumbersInExport, "record_grouping_" + profileName + "_ccontrol_numbers_to_ignore");
+			}
+			if (recordNumbersToIndex.size() > 0) {
+				writeExistingRecordsFile(configIni, recordNumbersToIndex, "record_grouping_" + profileName + "_bibs_to_index");
+			}
+			if (marcRecordsWritten.size() > 0) {
+				writeExistingRecordsFile(configIni, marcRecordsWritten, "record_grouping_" + profileName + "_new_bibs_written");
+			}
+			if (marcRecordsOverwritten.size() > 0) {
+				writeExistingRecordsFile(configIni, marcRecordsOverwritten, "record_grouping_" + profileName + "_changed_bibs_written");
+			}
 		}
 	}
 
-	private static void loadExistingMarcFiles(File individualMarcPath, HashSet<String> existingFiles) {
+	/*private static void loadExistingMarcFiles(File individualMarcPath, HashSet<String> existingFiles) {
 		File[] subFiles = individualMarcPath.listFiles();
 		if (subFiles != null){
 			for (File curFile : subFiles){
@@ -1255,13 +1281,16 @@ public class RecordGrouperMain {
 				}
 			}
 		}
-	}
+	}*/
 
-	private static int groupOverDriveRecords(Ini configIni, Connection econtentConnection, RecordGroupingProcessor recordGroupingProcessor, boolean explodeMarcsOnly) {
+	private static int groupOverDriveRecords(Ini configIni, Connection vufindConn, Connection econtentConnection, RecordGroupingProcessor recordGroupingProcessor, boolean explodeMarcsOnly) {
 		if (explodeMarcsOnly){
 			//Nothing to do since we don't have marc records to process
 			return 0;
 		}
+		loadIlsChecksums(vufindConn, "overdrive");
+		loadExistingPrimaryIdentifiers(vufindConn, "overdrive");
+
 		int numRecordsProcessed = 0;
 		try{
 			PreparedStatement overDriveRecordsStmt;
@@ -1337,24 +1366,10 @@ public class RecordGrouperMain {
 			overDriveRecordRS.close();
 
 			//This is no longer needed because we do cleanup differently now (get a list of everything in the database and then cleanup anything that isn't in the API anymore
-			if (!fullRegrouping){
-				/*PreparedStatement deletedRecordStmt;
-				if (lastGroupingTime == null || fullRegroupingNoClear){
-					deletedRecordStmt = econtentConnection.prepareStatement("SELECT overdriveId FROM overdrive_api_products WHERE deleted = 1",  ResultSet.TYPE_FORWARD_ONLY, ResultSet.CONCUR_READ_ONLY);
-				}else{
-					deletedRecordStmt = econtentConnection.prepareStatement("SELECT overdriveId FROM overdrive_api_products WHERE deleted = 1 and dateDeleted >= ?",  ResultSet.TYPE_FORWARD_ONLY, ResultSet.CONCUR_READ_ONLY);
-					deletedRecordStmt.setLong(1, lastGroupingTime);
-				}
-				ResultSet recordsToDelete = deletedRecordStmt.executeQuery();
-				while (recordsToDelete.next()){
-					RecordIdentifier primaryIdentifier = new RecordIdentifier();
-					String overdriveId = recordsToDelete.getString("overdriveId");
-					primaryIdentifier.setValue("overdrive", overdriveId);
-					recordGroupingProcessor.deletePrimaryIdentifier(primaryIdentifier);
-				}*/
-			}else{
+			if (fullRegrouping){
 				writeExistingRecordsFile(configIni, recordNumbersInExport, "record_grouping_overdrive_records_in_export");
 			}
+			removeDeletedRecords("overdrive");
 			logger.info("Finished grouping " + numRecordsProcessed + " records from overdrive ");
 		}catch (Exception e){
 			System.out.println("Error loading OverDrive records: " + e.toString());
@@ -1365,7 +1380,7 @@ public class RecordGrouperMain {
 
 	private static SimpleDateFormat oo8DateFormat = new SimpleDateFormat("yyMMdd");
 	private static SimpleDateFormat oo5DateFormat = new SimpleDateFormat("yyyyMMdd");
-	private static boolean writeIndividualMarc(HashSet<String> existingMarcFiles, IndexingProfile indexingProfile, Record marcRecord, String recordNumber, TreeSet<String> marcRecordsWritten, TreeSet<String> marcRecordsOverwritten) {
+	private static boolean writeIndividualMarc(IndexingProfile indexingProfile, Record marcRecord, String recordNumber, TreeSet<String> marcRecordsWritten, TreeSet<String> marcRecordsOverwritten) {
 		boolean marcRecordUpToDate = false;
 		//Copy the record to the individual marc path
 		if (recordNumber != null){
@@ -1378,7 +1393,7 @@ public class RecordGrouperMain {
 			//Check to see if the record needs to be written before writing it.
 			if (!fullRegrouping){
 				boolean checksumUpToDate = existingChecksum != null && existingChecksum.equals(checksum);
-				boolean fileExists = checkIfIndividualMarcFileExists(existingMarcFiles, true, individualFile);
+				boolean fileExists = checkIfIndividualMarcFileExists(true, individualFile);
 				marcRecordUpToDate = fileExists && checksumUpToDate;
 				if (!fileExists){
 					marcRecordsWritten.add(recordNumber);
@@ -1474,12 +1489,10 @@ public class RecordGrouperMain {
 		}
 	}
 
-	private static boolean checkIfIndividualMarcFileExists(HashSet<String> existingMarcFiles, Boolean marcRecordUpToDate, File individualFile) {
-		String filename = individualFile.getName();
-		if (!existingMarcFiles.contains(filename)){
+	private static boolean checkIfIndividualMarcFileExists(Boolean marcRecordUpToDate, File individualFile) {
+		if (!individualFile.exists()){
 			marcRecordUpToDate = false;
 		}
-		existingMarcFiles.remove(filename);
 		return marcRecordUpToDate;
 	}
